+++ src/properties/hirshfeld/CMakeLists.txt
@@ -1,4 +1,5 @@
 target_sources(mrchem PRIVATE
     ${CMAKE_CURRENT_SOURCE_DIR}/PolyInterpolator.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/logsumexp.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/HirshfeldInterpolator.cpp
     )
