+++ src/properties/hirshfeld/HirshfeldInterpolator.h
@@ -1,57 +1,38 @@
+#pragma once
+
 #include <vector>
 #include <fstream>
 #include <Eigen/Dense>
+#include <Eigen/Core>
 #include <string>
-#include <iostream>
-#include <filesystem>
+#include <memory>
 #include "PolyInterpolator.h"
 
-void readAtomicDensity(const std::string path, Eigen::VectorXd &rGrid, Eigen::VectorXd &rhoGrid){
-    std::vector<double> r, rho;
-    std::ifstream file(path);
-    std::string line;
-    double r_, rho_;
-    while (std::getline(file, line)) {
-        std::istringstream iss(line);
-        iss >> r_ >> rho_;
-        r.push_back(r_);
-        rho.push_back(rho_);
-    }
-    file.close();
-    rGrid = Eigen::Map<Eigen::VectorXd>(r.data(), r.size());
-    rhoGrid = Eigen::Map<Eigen::VectorXd>(rho.data(), rho.size());
-}
+/**
+ * @brief Reads atomic density data from a file
+ * @param path The path to the file containing the data
+ * @param rGrid The grid for the radial distances
+ * @param rhoGrid The grid for the atomic density values
+ */
+void readAtomicDensity(const std::string path, Eigen::VectorXd &rGrid, Eigen::VectorXd &rhoGrid);
 
 class HirshfeldRadInterpolater {
 
-    public:
+public:
     /**
      * @brief Construct a new Rad Interpolater object
      * @param element The element for which the ZORA potential is to be interpolated
      * @param data_dir The directory containing the ZORA potential data
     */
-    RadInterpolater(const std::string element, std::string data_dir){
-        Eigen::VectorXd rGrid;
-        Eigen::VectorXd rhoGrid;
+    HirshfeldRadInterpolater(const std::string element, std::string data_dir);
 
+    /**
+     * @brief Evaluate the interpolated function at a given point
+     * @param r The point at which to evaluate
+     * @return The interpolated value at the given point
+    */
+    double evalf(const double &r) const;
 
-        this->mode = mode;
-        std::string filename = data_dir + '/' + element + ".dat";
-
-        readAtomicDensity(filename, rGrid, rhoGrid);
-
-        rhoGrid = Eigen::log(rhoGrid); // todo multiply by 4pi*r^2 or some other magic number
-
-        lnRho = std::make_shared<PolyInterpolator>(rGrid, rhoGrid);
-
-    }
-
-    double evalf(const double &r) const {
-        double y, yp;
-        return lnRho->eval(r, y);
-    }
-
-    protected:
+protected:
     std::shared_ptr<PolyInterpolator> lnRho = nullptr;
-
 };
